+++ woorepie/src/main/java/com/piehouse/woorepie/subscription/dto/response/GetSubscriptionSimpleResponse.java
@@ -31,11 +31,11 @@ public class GetSubscriptionSimpleResponse {
 
     private String estateImageUrl;
 
-    private Integer tokenAmount;
+    private Long tokenAmount;
 
-    private Integer estatePrice;
+    private Long estatePrice;
 
-    private Integer estateTokenPrice;
+    private Long estateTokenPrice;
 
     private BigDecimal dividendYield;
